Simplify user bids and asks
