--- src/Client/src/app/account/user-asks/user-ask-edit/user-ask-edit.component.html
@@ -1,58 +1,0 @@
-<form [formGroup]="form" (ngSubmit)="onSubmit()">
-    <h2 mat-dialog-title>Edit ask</h2>
-    <mat-dialog-content class="mat-typography">
-        <div class="row" *ngIf="!deleted">
-          <div class="mb-3 col-6 offset-3" *ngIf="!nosize">
-            <label class="form-label" for="size">Size</label>
-            <select class="form-select mb-2" formControlName="size">
-              <option *ngFor="let size of allowedSizes" [value]="size">{{size}}</option>
-            </select>
-            <span *ngIf="form.get('size').invalid && form.get('size').touched"
-              class="form-text text-danger mb-2">
-              <span *ngIf="form.get('size').hasError('required')">
-                Size is required!
-              </span>
-              <span *ngIf="form.get('size').hasError('minlength')">
-                Size must have minimum 2 character!
-              </span>
-              <span *ngIf="form.get('size').hasError('maxlength')">
-                Size must have maximum 50 characters!
-              </span>
-            </span>
-          </div>
-          <div class="mb-3 col-6 offset-3">
-            <label class="form-label" for="price">Price</label>
-            <input formControlName="price" type="text" id="price" class="form-control" />
-            <span *ngIf="form.get('price').invalid && form.get('price').touched"
-              class="form-text text-danger mb-2">
-              <span *ngIf="form.get('price').hasError('required')">
-                Price is required!
-              </span>
-              <span *ngIf="form.get('price').hasError('minlength')">
-                Price must have minimum 2 character!
-              </span>
-              <span *ngIf="form.get('price').hasError('maxlength')">
-                Price must have maximum 50 characters!
-              </span>
-              <span *ngIf="form.get('price').hasError('min')">
-                Price must be at least $0.01!
-              </span>
-            </span>
-          </div>
-          <div class="alert alert-danger" role="alert" *ngIf="error === true">
-            Something went wrong
-          </div>
-        </div>
-        <div class="row" *ngIf="deleted">
-          <div class="alert alert-danger" role="alert">
-            Bid deleted
-          </div>
-        </div>
-    </mat-dialog-content>
-    <mat-dialog-actions align="end">
-      <button type="button" mat-button color="warn" (click)="onDelete()" *ngIf="!deleted">Delete ask</button>
-      <button type="button" mat-button mat-dialog-close>Exit</button>
-      <button type="submit" mat-button [disabled]="!form.valid" *ngIf="!deleted">Save</button>
-    </mat-dialog-actions>
-    </form>
-
