+++ src/Client/src/app/account/user-bids/user-bids.component.css
@@ -1,8 +1,1 @@
-/* table {
-    width: 100%;
-  } */
 
-  .mat-row:hover {
-    background: #f5f5f5;
-    cursor: pointer;
-  }
