+++ src/Client/src/app/account/user-bids/user-bids.component.html
@@ -1,64 +1,3 @@
-<div class="mt-5">
-  <app-loading-spinner *ngIf="loading && !error"></app-loading-spinner>
-  <div class="alert alert-danger" role="alert" *ngIf="error && !loading">
-    {{ error }}
-  </div>
-</div>
-<div class="container mt-3" *ngIf="loading === false">
-  <div class="row">
-    <div class="col-12 my-3">
-      <h1 class="text-center">
-        Bids
-      </h1>
-    </div>
-  </div>
-  <div class="row">
-    <table mat-table [dataSource]="dataSource" class="mat-elevation-z8">
-
-      <ng-container matColumnDef="id">
-        <th mat-header-cell *matHeaderCellDef></th>
-        <td mat-cell *matCellDef="let element"></td>
-      </ng-container>
-
-      <ng-container matColumnDef="name">
-        <th mat-header-cell *matHeaderCellDef> Name </th>
-        <td mat-cell *matCellDef="let element"> {{element.name | titlecase}} </td>
-      </ng-container>
-
-      <ng-container matColumnDef="price">
-        <th mat-header-cell *matHeaderCellDef> Price </th>
-        <td mat-cell *matCellDef="let element"> {{element.price | currency}} </td>
-      </ng-container>
-
-      <ng-container matColumnDef="fee">
-        <th mat-header-cell *matHeaderCellDef> Fee </th>
-        <td mat-cell *matCellDef="let element"> {{element.fee | currency}} </td>
-      </ng-container>
-
-      <ng-container matColumnDef="size">
-        <th mat-header-cell *matHeaderCellDef> Size </th>
-        <td mat-cell *matCellDef="let element"> {{element.size | titlecase}} </td>
-      </ng-container>
-
-      <ng-container matColumnDef="lowestAsk">
-        <th mat-header-cell *matHeaderCellDef> Lowest Ask </th>
-        <td mat-cell *matCellDef="let element"> {{element.lowestAsk}} </td>
-      </ng-container>
-
-      <ng-container matColumnDef="highestBid">
-        <th mat-header-cell *matHeaderCellDef> Highest Bid </th>
-        <td mat-cell *matCellDef="let element"> {{element.highestBid}} </td>
-      </ng-container>
-
-      <ng-container matColumnDef="expires">
-        <th mat-header-cell *matHeaderCellDef> Expires </th>
-        <td mat-cell *matCellDef="let element"> {{element.expires | date}} </td>
-      </ng-container>
-
-      <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-      <tr mat-row *matRowDef="let row; columns: displayedColumns;" (click)="onEdit(row)"></tr>
-    </table>
-    <mat-paginator hidePageSize [pageSize]="10" [pageIndex]="paginationData.pageIndex-1" showFirstLastButtons [length]="paginationData.totalCount" (page)="pageEvent = changePage($event)">
-    </mat-paginator>
-  </div>
+<div class="row">
+  <app-user-orders-table [type]="type"></app-user-orders-table>
 </div>
