+++ src/Client/src/app/account/user-orders-table/order-edit/update-order.ts
@@ -1,3 +1,3 @@
-export class UpdateAsk {
+export class UpdateOrder {
   constructor(public id: string, public size: string, public price: string) {}
 }
