+++ src/Client/src/app/account/user-orders-table/user-orders-table.component.css
@@ -1,0 +1,8 @@
+table {
+    width: 100%;
+  }
+
+.mat-row:hover {
+    background: #f5f5f5;
+    cursor: pointer;
+}
