+++ src/Client/src/app/app.module.ts
@@ -49,10 +49,10 @@ import { MatTableModule } from '@angular/material/table';
 import { MatPaginatorModule } from '@angular/material/paginator';
 import { AsksComponent } from './shared/asks/asks.component';
 import { BidsComponent } from './shared/bids/bids.component';
-import { UserBidEditComponent } from './account/user-bids/user-bid-edit/user-bid-edit.component';
 import { UserBidsComponent } from './account/user-bids/user-bids.component';
 import { UserAsksComponent } from './account/user-asks/user-asks.component';
-import { UserAskEditComponent } from './account/user-asks/user-ask-edit/user-ask-edit.component';
+import { UserOrdersTableComponent } from './account/user-orders-table/user-orders-table.component';
+import { OrderEditComponent } from './account/user-orders-table/order-edit/order-edit.component';
 
 @NgModule({
   declarations: [
@@ -91,9 +91,9 @@ import { UserAskEditComponent } from './account/user-asks/user-ask-edit/user-ask
     SellerLevelComponent,
     SellerLevelProgressComponent,
     UserBidsComponent,
-    UserBidEditComponent,
     UserAsksComponent,
-    UserAskEditComponent
+    UserOrdersTableComponent,
+    OrderEditComponent
   ],
   imports: [
     BrowserModule,
