--- src/Client/src/app/shared/bids/update-bid.model.ts
@@ -1,3 +1,0 @@
-export class UpdateBid {
-  constructor(public id: string, public size: string, public price: string) {}
-}
